+++ examples/rust/src/bin/client.rs
@@ -3,10 +3,11 @@ use {
     clap::{Parser, Subcommand, ValueEnum},
     futures::{future::TryFutureExt, sink::SinkExt, stream::StreamExt},
     log::{error, info},
-    solana_sdk::pubkey::Pubkey,
+    solana_sdk::{pubkey::Pubkey, signature::Signature},
+    solana_transaction_status::{EncodedTransactionWithStatusMeta, UiTransactionEncoding},
     std::{
         collections::HashMap,
-        env,
+        env, fmt,
         sync::{Arc, Mutex},
         time::Duration,
     },
@@ -20,7 +21,7 @@ use {
             SubscribeRequestFilterAccountsFilter, SubscribeRequestFilterAccountsFilterMemcmp,
             SubscribeRequestFilterBlocks, SubscribeRequestFilterBlocksMeta,
             SubscribeRequestFilterEntry, SubscribeRequestFilterSlots,
-            SubscribeRequestFilterTransactions, SubscribeUpdateAccount,
+            SubscribeRequestFilterTransactions, SubscribeUpdateAccount, SubscribeUpdateTransaction,
         },
         tonic::service::Interceptor,
     },
@@ -353,6 +354,48 @@ impl From<SubscribeUpdateAccount> for AccountPretty {
     }
 }
 
+#[allow(dead_code)]
+pub struct TransactionPretty {
+    slot: u64,
+    signature: Signature,
+    is_vote: bool,
+    tx: EncodedTransactionWithStatusMeta,
+}
+
+impl fmt::Debug for TransactionPretty {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        struct TxWrap<'a>(&'a EncodedTransactionWithStatusMeta);
+        impl<'a> fmt::Debug for TxWrap<'a> {
+            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+                let serialized = serde_json::to_string(self.0).expect("failed to serialize");
+                fmt::Display::fmt(&serialized, f)
+            }
+        }
+
+        f.debug_struct("TransactionPretty")
+            .field("slot", &self.slot)
+            .field("signature", &self.signature)
+            .field("is_vote", &self.is_vote)
+            .field("tx", &TxWrap(&self.tx))
+            .finish()
+    }
+}
+
+impl From<SubscribeUpdateTransaction> for TransactionPretty {
+    fn from(SubscribeUpdateTransaction { transaction, slot }: SubscribeUpdateTransaction) -> Self {
+        let tx = transaction.expect("should be defined");
+        Self {
+            slot,
+            signature: Signature::try_from(tx.signature.as_slice()).expect("valid signature"),
+            is_vote: tx.is_vote,
+            tx: yellowstone_grpc_proto::convert_from::create_tx_with_meta(tx)
+                .expect("valid tx with meta")
+                .encode(UiTransactionEncoding::Base64, Some(u8::MAX), true)
+                .expect("failed to encode"),
+        }
+    }
+}
+
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     env::set_var(
@@ -482,6 +525,14 @@ async fn geyser_subscribe(
                         info!(
                             "new account update: filters {:?}, account: {:#?}",
                             msg.filters, account
+                        );
+                        continue;
+                    }
+                    Some(UpdateOneof::Transaction(tx)) => {
+                        let tx: TransactionPretty = tx.into();
+                        info!(
+                            "new transaction update: filters {:?}, transaction: {:#?}",
+                            msg.filters, tx
                         );
                         continue;
                     }